#4 Created the MyWishListFragment which will host the packages that the user wants to have saved on its own little list for future travels.
The items will look the exact same way as the ones in the PackageFragment, therefore it makes sense to leave the same packages_items.xml layout for the items and since the fragment only hosts a recyclerView (maybe for now) it also makes sense to inflate it with the fragment_packages.xml layout, which only carries a recyclerView.

The MyWishListFragment is almost functionally completed, it is only missing the way of populating it, which is still waiting for the TripViewFragment to set up the add to wish list button.